backend/requirements.txt
aiofiles==25.1.0
asyncpg==0.31.0
bcrypt==4.1.3
certifi==2026.2.25
fastapi==0.110.1
httpx==0.28.1
openpyxl==3.1.5
PyJWT==2.12.1
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
starlette==0.37.2
uvicorn[standard]==0.25.0
